be_dev/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
django-cors-headers==3.5.0
gunicorn==20.0.4
requests==2.24.0
PyJWT==2.0.1
djangorestframework-simplejwt==4.6.0
django-filter==2.4.0